fix:#41 - code review reflected
