+++ src/main/java/com/req2res/actionarybe/global/exception/ErrorCode.java
@@ -57,9 +57,13 @@ public enum ErrorCode {
 
 	INVALID_TOKEN(HttpStatus.UNAUTHORIZED, "유효하지 않은 토큰입니다."),
 	EXPIRED_TOKEN(HttpStatus.UNAUTHORIZED, "토큰이 만료되었습니다."),
-	MISSING_TOKEN(HttpStatus.UNAUTHORIZED, "토큰이 존재하지 않습니다.");
+	MISSING_TOKEN(HttpStatus.UNAUTHORIZED, "토큰이 존재하지 않습니다."),
 
-	private final HttpStatus status;
+    // badge
+    BADGE_NOT_FOUND(HttpStatus.INTERNAL_SERVER_ERROR, "뱃지가 존재하지 않습니다.");
+
+
+    private final HttpStatus status;
 	private final String message;
 
 }
